fix: relocate type files to a types folder
