+++ packages/components/src/components/hds/alert/title.ts
@@ -4,6 +4,6 @@
  */
 
 import Component from '@glimmer/component';
-import type { HdsAlertTitleSignature } from './types.ts';
+import type { HdsAlertTitleSignature } from '../../../types/components/alert.ts';
 
 export default class HdsAlertTitleComponent extends Component<HdsAlertTitleSignature> {}
