+++ packages/components/src/components/hds/text/display.ts
@@ -5,14 +5,17 @@
 
 import Component from '@glimmer/component';
 import { assert } from '@ember/debug';
-import { HdsTextSizeValues, HdsTextWeightValues } from './types.ts';
+import {
+  HdsTextSizeValues,
+  HdsTextWeightValues,
+} from '../../../types/components/text.ts';
 import type {
   HdsTextAligns,
   HdsTextColors,
   HdsTextSizes,
   HdsTextTags,
   HdsTextWeights,
-} from './types.ts';
+} from '../../../types/components/text.ts';
 
 // notice: only some combinations of size + font-weight are allowed (per design specs)
 // see: https://www.figma.com/file/oQsMzMMnynfPWpMEt91OpH/HDS-Product---Foundations?node-id=1262%3A9192
