+++ packages/components/src/components/hds/yield/index.ts
@@ -4,6 +4,6 @@
  */
 
 import Component from '@glimmer/component';
-import type { HdsYieldSignature } from './types';
+import type { HdsYieldSignature } from '../../../types/components/yield';
 
 export default class HdsYieldComponent extends Component<HdsYieldSignature> {}
