+++ packages/components/src/types/components/text.ts
